% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createDesignMatrix.R
\name{createDesignMatrix}
\alias{createDesignMatrix}
\title{Create design matrix.}
\usage{
createDesignMatrix(group, nexons)
}
\arguments{
\item{group}{Vector containing group identifiers for each sample.}

\item{nexons}{Number of exons in the gene.}
}
\value{
Returns a design matrix in the format required by the functions in the
  \code{regsplice} package.
}
\description{
Creates a gene-level design matrix in the format required by the \code{regsplice}
package.
}
\details{
Creates a design matrix for a single gene, given the number of exons and group
membership for each sample. This function will typically be called once for every gene
in a data set.

Notes:

\itemize{
  \item No main effect is included for the group factor. The group main effect is not
  required since it is absorbed within the sample main effects.
  \item No intercept is included, since it is simpler to let the \code{glmnet}
  functions add it back later.
}
}
\examples{
group <- rep(c(0, 1), each = 3)
nexons <- 10
createDesignMatrix(group, nexons)
}

