% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_exons.R
\name{filter_exons}
\alias{filter_exons}
\title{Filter low-count exons.}
\usage{
filter_exons(Y, n1 = 4, n2 = 3)
}
\arguments{
\item{Y}{RNA-seq read counts for multiple genes (list of data frames or matrices).
Created using the data preparation function \code{\link{prepare_data}}.}

\item{n1}{Filtering parameter: minimum number of reads per exon, summed across all 
biological samples. Default is 4.}

\item{n2}{Filtering parameter: minimum number of reads for a single biological sample
per exon, i.e. at least one sample must have this number of reads. Default is 3.}
}
\value{
Returns a list of data frames, where each data frame in the list contains the 
  RNA-seq read counts for one gene. Low-count exons and any remaining single-exon
  genes have been removed.
}
\description{
Filter low-count exons from RNA-seq read count data.
}
\details{
Filters low-count exons from an RNA-seq read count data set prepared with the
\code{\link{prepare_data}} function, i.e. in the format required by other functions in
the \code{regsplice} pipeline.

The arguments \code{n1} and \code{n2} control the amount of filtering. Exons that meet
the filtering conditions specified by \code{n1} and \code{n2} are kept. Default values
are provided; however these may be sub-optimal depending on the experimental design.

Any single-exon genes that remain after exons have been removed during the filtering 
step are also removed (since differential splicing requires multiple exons). The
output is in the same format as that from \code{\link{prepare_data}}.
}
\examples{
file_counts <- system.file("extdata/counts.txt", package = "regsplice")
data <- read.table(file_counts, header = TRUE, sep = "\\t", stringsAsFactors = FALSE)
counts <- data[, 2:7]
gene <- sapply(strsplit(data$exon, ":"), function(s) s[[1]])

Y <- prepare_data(counts, gene)
Y <- filter_exons(Y)

}
\seealso{
Other prepare_data filter_exons: \code{\link{prepare_data}}
}

