% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{filter_single_exons}
\alias{filter_single_exons}
\title{Filter genes containing a single exon.}
\usage{
filter_single_exons(Y)
}
\arguments{
\item{Y}{RNA-seq count table in the format prepared by \code{\link{split_genes}}; i.e.
a list of data frames, one for each gene.}
}
\value{
Y Returns a list of data frames, where each data frame contains the RNA-seq 
  counts for one gene, and genes containing only a single exon have been removed.
}
\description{
Filter genes containing only a single exon, from a data set prepared with
\code{\link{split_genes}}.
}
\details{
Input data should be in the format prepared by the \code{\link{split_genes}} function,
i.e. a list of data frames, where each data frame contains RNA-seq counts for one
gene.
}
\examples{
counts <- rbind(matrix(1, nrow = 3, ncol = 4), 
                matrix(1, nrow = 2, ncol = 4), 
                matrix(1, nrow = 1, ncol = 4), 
                matrix(0, nrow = 5, ncol = 4), 
                matrix(1, nrow = 2, ncol = 4))
gene <- paste0("gene", rep(1:5, times = c(3, 2, 1, 5, 2)))
Y <- split_genes(counts, gene)
names(Y)
length(Y)
Y <- filter_single_exons(Y)
names(Y)
length(Y)

}
\seealso{
Other split_genes filter_zeros filter_single_exons: \code{\link{filter_zeros}}
}

