% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{filter_zeros}
\alias{filter_zeros}
\title{Filter genes with zero counts.}
\usage{
filter_zeros(Y)
}
\arguments{
\item{Y}{RNA-seq count table in the format prepared by \code{\link{split_genes}}; i.e.
a list of data frames, one for each gene.}
}
\value{
Y Returns a list of data frames, where each data frame contains the RNA-seq 
  counts for one gene, and genes with zero total counts have been removed.
}
\description{
Filter genes with zero counts from a data set prepared with \code{\link{split_genes}}.
}
\details{
Input data should be in the format prepared by the \code{\link{split_genes}} function,
i.e. a list of data frames, where each data frame contains RNA-seq counts for one
gene.
}
\examples{
counts <- rbind(matrix(1, nrow = 3, ncol = 4), 
                matrix(1, nrow = 2, ncol = 4), 
                matrix(0, nrow = 4, ncol = 4), 
                matrix(1, nrow = 2, ncol = 4))
gene <- paste0("gene", rep(1:4, times = c(3, 2, 4, 2)))
Y <- split_genes(counts, gene)
names(Y)
length(Y)
Y <- filter_zeros(Y)
names(Y)
length(Y)

}
\seealso{
Other split_genes filter_zeros filter_single_exons: \code{\link{filter_single_exons}}
}

