% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\title{Fit full models with all interaction terms.}
\usage{
fitGLM(X_genes = NULL, Y_genes, weights_genes = NULL, group = NULL,
  nexons_genes = NULL, return_fitted = FALSE, ncores = 1, ...)
}
\arguments{
\item{X_genes}{List of design matrices. Individual design matrices in the required
format can be created with the function \code{\link{createDesignMatrix}}. Note that
column names are used to identify interaction terms (see above). Can be set to
\code{NULL}, in which case \code{group} and \code{nexons_genes} must be provided.}

\item{Y_genes}{List of expression data matrices. Assumed to be on a continuous scale,
for example transformed using the \code{voom} function from the \code{limma}
package.}

\item{weights_genes}{Optional list of weights matrices, for example generated by
\code{voom}.}

\item{group}{Optional vector of group membership identifiers for each sample. Must be
provided if \code{X_genes = NULL}, so that \code{\link{createDesignMatrix}} can
be called to generate design matrices.}

\item{nexons_genes}{Optional vector containing number of exons for each gene. Must be
provided if \code{X_genes = NULL}, so that \code{\link{createDesignMatrix}} can
be called to generate design matrices.}

\item{return_fitted}{Logical; whether to return fitted model objects from
\code{glm}.}

\item{ncores}{Number of cores for parallel evaluation. Default is 1. Note that on
Windows systems, the code will always run on a single core.}

\item{...}{Other arguments to pass to \code{glm}.}
}
\value{
Returns a list containing:
\itemize{
  fit_genes: (optional) fitted \code{glm} model objects for each gene
  dev_genes: deviance of fitted model for each gene
  df_genes: degrees of freedom of fitted model for each gene
}
}
\description{
Parallelized function to fit full GLMs (with all interaction terms) for multiple genes
using \code{glm}.
}
\details{
Fits full GLMs containing all interaction terms for each gene. Model fitting is done
with \code{glm}. These models can be used together with the nested null models (no
interaction terms) from \code{\link{fitNullModels}} to calculate likelihood ratio
tests with the function \code{\link{lrTest}}.

The full models are also used to replace fitted models containing zero interaction
terms if the option \code{when_null_selected = "GLM"} is used in \code{\link{lrTest}}
(see documentation for \code{\link{lrTest}} for more details).

The full models contain exon main effects, sample main effects, and all exon:group
interaction terms. This is similar to the approach used in \code{voom-diffSlice}
(function \code{diffSplice} in the \code{limma} package).
}
\examples{
set.seed(1)
group <- rep(c(0, 1), each=3)
nexons <- 8
X <- createDesignMatrix(group, nexons)
Y <- rnorm(nrow(X), mean=2, sd=1)
ix <- c(7, 8) + ( 8 * rep(0:5, each=2) )
Y[ix] <- Y[ix] + 1
fitGLM(X, Y)
}
\seealso{
\code{\link{fitRegModelSingle}} \code{\link{fitGLMSingle}}
  \code{\link{fitNullModelSingle}} \code{\link{lrTest}} \code{\link[stats]{glm}}

Other X.code..link.fitRegModel....code..link.fitGLM....code..link.fitNullModel..: \code{\link{fitNullModel}};
  \code{\link{fitRegModel}}
}

