% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLMSingle}
\alias{fitGLMSingle}
\title{Fit full model with all interaction terms (single gene).}
\usage{
fitGLMSingle(X = NULL, Y, weights = NULL, group = NULL, nexons = NULL,
  ...)
}
\description{
Internal function to fit full GLM (with all interaction terms) for a single gene.
}
\details{
In most cases the user will use the external function \code{\link{fitGLM}}, which can
fit models for multiple genes using parallelization. \code{fitGLM} calls
\code{fitGLMSingle} internally.

See documentation for \code{\link{fitGLM}} for more details.
}
\seealso{
\code{\link{fitRegModel}} \code{\link{fitGLM}} \code{\link{fitNullModel}}
  \code{\link{lrTest}} \code{\link[stats]{glm}}

Other fitRegModelSingle.fitGLMSingle.fitNullModelSingle: \code{\link{fitNullModelSingle}};
  \code{\link{fitRegModelSingle}}
}

