% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitNullModel.R
\name{fitNullModelSingle}
\alias{fitNullModelSingle}
\title{Fit null model (single gene).}
\usage{
fitNullModelSingle(X = NULL, Y, weights = NULL, group = NULL,
  nexons = NULL, ...)
}
\description{
Internal function to fit null model (no interaction terms) for a single gene.
}
\details{
In most cases the user will use the external function \code{\link{fitNullModel}},
which can fit null models for multiple genes using parallelization.
\code{fitNullModel} calls \code{fitNullModelSingle} internally.

See documentation for \code{\link{fitNullModel}} for more details.
}
\seealso{
\code{\link{fitRegModel}} \code{\link{fitGLM}} \code{\link{fitNullModel}}
  \code{\link{lrTest}} \code{\link[stats]{glm}}

Other fitRegModelSingle.fitGLMSingle.fitNullModelSingle: \code{\link{fitGLMSingle}};
  \code{\link{fitRegModelSingle}}
}

