% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitParallel.R
\name{fitParallel}
\alias{fitParallel}
\title{Fit multiple genes (parallelized).}
\usage{
fitParallel(fitting_function = c("regularized", "GLM", "null"),
  X_genes = NULL, Y_genes, weights_genes = NULL, group = NULL,
  nexons_genes = NULL, alpha = 1, lambda_choice = c("lambda.min",
  "lambda.1se"), return_fitted = FALSE, ncores = 1, ...)
}
\arguments{
\item{fitting_function}{Which single-gene fitting function to use. Choices are 
"regularized" (\code{\link{fitRegModelSingle}}), "GLM" (\code{\link{fitGLMSingle}}),
or "null" (\code{\link{fitNullSingle}}). All other arguments (except
\code{fitting_function} and \code{ncores}) will be passed to the single-gene fitting
function. See documentation for \code{\link{fitRegModel}}, \code{\link{fitGLM}}, and
\code{\link{fitNull}} for information about the other arguments.}

\item{ncores}{Number of cores for parallel evaluation. Default is 1. Note that on
Windows systems, the code will always run on a single core.}
}
\value{
Returns a list containing:
\itemize{
\item fit_genes: (list) fitted model objects
\item dev_genes: (vector) deviance of fitted models
\item df_genes: (vector) degrees of freedom of fitted models
}
}
\description{
Internal function to apply model fitting functions to multiple genes using
parallelization.
}
\details{
The internal model fitting functions (\code{\link{fitRegModelSingle}}, 
\code{\link{fitGLMSingle}}, and \code{\link{fitNullModelSingle}}) are designed for 
fitting single genes only. \code{fitParallel} applies these functions to multiple
genes, using parallelization to speed up runtime.

In most cases, the user will access only the external functions
\code{\link{fitRegModel}}, \code{\link{fitGLM}}, and \code{\link{fitNull}}. See
documentation for those functions for more details.
}
\seealso{
\code{\link{fitRegModel}} \code{\link{fitGLM}} \code{\link{fitNullModel}} 
  \code{\link{fitRegModelSingle}} \code{\link{fitGLMSingle}}
  \code{\link{fitNullModelSingle}}
}

