% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitRegModel.R
\name{fitRegModelSingle}
\alias{fitRegModelSingle}
\title{Fit regularized model (single gene).}
\usage{
fitRegModelSingle(X = NULL, Y, weights = NULL, group = NULL,
  nexons = NULL, alpha = 1, lambda_choice = c("lambda.min", "lambda.1se"),
  ...)
}
\description{
Internal function to fit regularized model for a single gene.
}
\details{
In most cases the user will use the external function \code{\link{fitRegModel}}, which
can fit regularized models for multiple genes using parallelization.
\code{fitRegModel} calls \code{fitRegModelSingle} internally.

See documentation for \code{\link{fitRegModel}} for more details.
}
\seealso{
\code{\link{fitRegModel}} \code{\link{fitGLM}} \code{\link{fitNullModel}}
  \code{\link{lrTest}} \code{\link[glmnet]{glmnet}} \code{\link[glmnet]{cv.glmnet}}

Other fitRegModelSingle.fitGLMSingle.fitNullModelSingle: \code{\link{fitGLMSingle}};
  \code{\link{fitNullModelSingle}}
}

