% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_tests.R
\name{LR_tests}
\alias{LR_tests}
\title{Calculate likelihood ratio tests.}
\usage{
LR_tests(fitted_models_reg, fitted_models_GLM = NULL, fitted_models_null,
  when_null_selected = c("ones", "GLM", "NA"))
}
\arguments{
\item{fitted_models_reg}{Regularized (lasso) fitted models (output from
\code{\link{fit_reg}}.}

\item{fitted_models_GLM}{Full GLM fitted models (output from \code{\link{fit_GLM}}).
Not required if \code{when_null_selected = "ones" or "NA"}. Default is \code{NULL}.}

\item{fitted_models_null}{Null models (output from \code{\link{fit_null}}).}

\item{when_null_selected}{Which option to use for genes where the lasso model selects 
zero interaction terms, i.e. identical to the null model. Options are \code{"ones"},
\code{"GLM"}, and \code{"NA"}. Default is \code{"ones"}. See below for details.}
}
\value{
Returns a list containing:
\itemize{
\item p_vals: p-values
\item p_adj: p-values adjusted for multiple testing (Benjamini-Hochberg false
  discovery rates)
\item LR_stats: likelihood ratio statistics
\item df_tests: degrees of freedom of the likelihood ratio tests
}
}
\description{
Calculate likelihood ratio tests between fitted models and null models.
}
\details{
The regularized (lasso) fitted models contain an optimal subset of exon:condition 
interaction terms; the GLM fitted models contain all exon:condition interaction terms;
and the null models do not contain any interaction terms. The null models are
therefore nested within the fitted models.

The likelihood ratio (LR) tests compare the fitted models against the nested null
models.

If the regularized (lasso) model contains at least one exon:condition interaction 
term, the LR test compares the lasso model against the null model. However, if the 
lasso model contains zero interaction terms, then the lasso and null models are 
identical, so the LR test cannot be calculated. The \code{when_null_selected} argument
lets the user choose what to do in these cases: either set p-values equal to 1 
(\code{when_null_selected = "ones"}); or calculate a LR test using the full GLM
containing all exon:condition interaction terms (\code{when_null_selected = "GLM"}),
which reduces power due to the larger number of terms, but allows the evidence for
differential exon usage among these genes to be distinguished. You can also return
\code{NA}s for these genes (\code{when_null_selected = "NA"}).

The default option is \code{when_null_selected = "ones"}. This simply calls all these
genes non-significant, which in most cases is sufficient since we are more interested
in genes with strong evidence for differential exon usage. However, if it is important
to rank the low-evidence genes in your data set, then use the \code{when_null_selected
= GLM} option.

If \code{when_null_selected = "ones"}, the full GLM fitted models are not required, so
you can set \code{fitted_models_GLM = NULL} (the default).
}
\examples{
condition <- rep(c(0, 1), each = 3)
n_exons <- 10
Y <- list(as.data.frame(matrix(sample(100:200, 60, replace = TRUE), nrow = 10)))
fitted_models_reg <- fit_reg(Y, condition)
fitted_models_GLM <- fit_GLM(Y, condition)
fitted_models_null <- fit_null(Y, condition)

LR_tests(fitted_models_reg = fitted_models_reg, 
         fitted_models_GLM = NULL, 
         fitted_models_null = fitted_models_null, 
         when_null_selected = "ones")

LR_tests(fitted_models_reg = fitted_models_reg, 
         fitted_models_GLM = fitted_models_GLM, 
         fitted_models_null = fitted_models_null, 
         when_null_selected = "GLM")

}
\seealso{
Other create_design_matrix fit_reg fit_GLM fit_null LR_tests: \code{\link{create_design_matrix}},
  \code{\link{fit_reg}}
}

