% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data.}
\usage{
prepare_data(counts, gene)
}
\arguments{
\item{counts}{RNA-seq counts (matrix or data frame). Each row is an exon, and each
column is a biological sample.}

\item{gene}{Vector of gene identifiers (character vector). Length is equal to the 
number of rows in \code{counts}.}
}
\value{
Returns a list of data frames, where each data frame contains the RNA-seq 
  counts for one gene.
}
\description{
Function to prepare data into the format required by subsequent functions in the
\emph{regsplice} pipeline.
}
\details{
Splits a matrix (or data frame) of RNA-seq counts into a list of data frames, where 
each data frame in the list contains the RNA-seq counts for one gene.

Gene identifiers are provided in the \code{gene} argument, which should contain one
entry for every exon in every gene; i.e. the length of \code{gene} should be equal to
the number of rows in \code{counts}, with repeated entries for for multiple exons in
each gene.
}
\examples{
counts <- matrix(sample(100:200, 7 * 4, replace = TRUE), nrow = 7, ncol = 4)
gene <- paste0("gene", rep(1:3, times = c(3, 2, 2)))
Y <- prepare_data(counts, gene)

file_counts <- system.file("extdata/counts.txt", package = "regsplice")
data <- read.table(file_counts, header = TRUE, sep = "\\t", stringsAsFactors = FALSE)
counts <- data[, 2:7]
gene <- sapply(strsplit(data$exon, ":"), function(s) s[[1]])
Y <- prepare_data(counts, gene)

}

