% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsplice_wrapper.R
\name{regsplice}
\alias{regsplice}
\title{Wrapper function to run regsplice.}
\usage{
regsplice(counts = NULL, gene_IDs = NULL, n_exons = NULL,
  condition = NULL, input_SummarizedExperiment = NULL,
  filter_zeros = TRUE, filter_low_counts = TRUE, filter_min_per_exon = 6,
  filter_min_per_sample = 3, normalize = TRUE, norm_method = "TMM",
  voom = TRUE, alpha = 1, lambda_choice = c("lambda.min", "lambda.1se"),
  when_null_selected = c("ones", "full", "NA"), n_cores_reg = NULL,
  n_cores_null = 1, n_cores_full = 1, seed = NULL, progress_bar = TRUE,
  ...)
}
\arguments{
\item{counts}{RNA-seq read counts or exon microarray intensities (matrix or data 
frame). Rows are exons, and columns are biological samples.}

\item{gene_IDs}{Vector of gene IDs (character vector). Length is equal to the number 
of genes.}

\item{n_exons}{Vector of exon lengths (numeric vector of integers), i.e. the number of
exon bins per gene. Length is equal to the number of genes.}

\item{condition}{Experimental condition for each biological sample (character or 
numeric vector, or factor).}

\item{input_SummarizedExperiment}{Optional alternative input argument, for providing 
inputs as a \code{SummarizedExperiment} object (see details below). This may be 
useful when running \code{regsplice} as part of a pipeline. If \code{NULL} 
(default), all four main input arguments must be provided instead (\code{counts}, 
\code{gene_IDs}, \code{n_exons}, \code{condition}).}

\item{filter_zeros}{Whether to filter zero-count exon bins, using
\code{\link{filter_zeros}}. Default is TRUE. Set to FALSE for exon microarray data.}

\item{filter_low_counts}{Whether to filter low-count exon bins, using
\code{\link{filter_low_counts}}. Default is TRUE. Set to FALSE for exon microarray
data.}

\item{filter_min_per_exon}{Filtering parameter for low-count exon bins: minimum number
of reads per exon bin, summed across all biological samples. Default is 6. See
\code{\link{filter_low_counts}} for details.}

\item{filter_min_per_sample}{Filtering parameter for low-count exon bins: minimum
number of reads per biological sample; i.e. for each exon bin, at least one sample
must have this number of reads. Default is 3. See \code{\link{filter_low_counts}}
for details.}

\item{normalize}{Whether to calculate normalization factors, using 
\code{\link{run_normalization}}. Default is TRUE. If FALSE, non-normalized library 
sizes will be used. Set to FALSE for exon microarray data.}

\item{norm_method}{Normalization method to use. Options are \code{"TMM"}, 
\code{"RLE"}, \code{"upperquartile"}, and \code{"none"}. Default is \code{"TMM"}. 
See \code{\link{run_normalization}} for details.}

\item{voom}{Whether to calculate \code{limma-voom} transformation and weights, using 
\code{\link{run_voom}}. Default is TRUE. If FALSE, model fitting functions will use 
the raw input data (not recommended for count data) with exon bins weighted equally.
Set to FALSE for exon microarray data.}

\item{alpha}{Elastic net parameter \code{alpha} for \code{glmnet} model fitting 
functions. Must be between 0 (ridge regression) and 1 (lasso). Default is 1 (lasso).
See \code{glmnet} documentation for more details.}

\item{lambda_choice}{Parameter to select which optimal \code{lambda} value to choose 
from the \code{cv.glmnet} cross validation fit. Choices are "lambda.min" (model with
minimum cross-validated error) and "lambda.1se" (most regularized model with 
cross-validated error within one standard error of minimum). Default is 
"lambda.min". See \code{glmnet} documentation for more details.}

\item{when_null_selected}{Which option to use for genes where the lasso model selects 
zero interaction terms, i.e. identical to the null model. Options are \code{"ones"},
\code{"full"}, and \code{"NA"}. Default is \code{"ones"}. See \code{\link{LR_tests}}
for details.}

\item{n_cores_reg}{Number of processor cores for fitting regularized models. Default 
is 8, or the maximum available if less than 8.}

\item{n_cores_null}{Number of processor cores for fitting null models. Default is 1, 
since this function is already very fast.}

\item{n_cores_full}{Number of processor cores for fitting full models. Default is 1,
since this function is already very fast.}

\item{seed}{Random seed (integer). Default is NULL. Provide an integer value to set 
the random seed for reproducible results.}

\item{progress_bar}{Whether to display progress bar during model fitting (regularized 
models only). Default is TRUE.}

\item{...}{Other arguments to pass to \code{cv.glmnet}, \code{glmnet}, or \code{glm}.}
}
\value{
Returns a \code{\linkS4class{RegspliceResults}} object containing fitted model
  objects and likelihood ratio (LR) test results. The LR test results consist of the
  following entries for each gene:
\itemize{
\item p_vals: raw p-values
\item p_adj: multiple testing adjusted p-values (Benjamini-Hochberg false discovery 
rates, FDR)
\item LR_stats: likelihood ratio test statistics
\item df_tests: degrees of freedom of likelihood ratio tests
}
}
\description{
Wrapper function to run a complete \code{regsplice} analysis with one command.
}
\details{
This wrapper function runs the complete \code{regsplice} pipeline with a single 
command. It calls each of the individual functions within the package in sequence. You
can also run the individual functions separately, which provides additional 
flexibility and insight into the statistical methodology. See the vignette for a 
description of the individual functions and an example workflow.

Required inputs are \code{counts} (matrix or data frame of RNA-seq read counts or exon
microarray intensities), \code{gene_IDs} (vector of gene IDs), \code{n_exons} (vector
of exon lengths, i.e. number of exon bins per gene), and \code{condition} (vector of
experimental conditions for each biological sample).

Alternatively, the inputs can be provided as a \code{SummarizedExperiment} object, 
which will be parsed to extract each of these components. This may be useful when 
running \code{regsplice} as part of a pipeline together with other packages.

See the vignette for an example showing how to construct \code{gene_IDs} and
\code{n_exons} from a column of gene:exon IDs.

After running the analysis, a summary table of the results can be displayed with
\code{\link{summary_table}}.

Exon microarray intensities should be log2-transformed, which is usually done during 
pre-processing of microarray data. (RNA-seq counts will be transformed automatically 
using \code{limma-voom}; see \code{\link{run_voom}}.)

In addition, when using exon microarray data, the filtering, normalization, and
\code{voom} steps should be disabled with the respective arguments.

See \code{\link{filter_zeros}} and \code{\link{filter_low_counts}} for details about
filtering; \code{\link{run_normalization}} and \code{\link{run_voom}} for details
about calculation of normalization factors and \code{voom} transformation and weights;
\code{\link{create_design_matrix}} for details about the model design matrices; 
\code{\link{fit_reg_multiple}}, \code{\link{fit_null_multiple}}, or 
\code{\link{fit_full_multiple}} for details about the model fitting functions; and 
\code{\link{LR_tests}} for details about the likelihood ratio tests.
}
\examples{
file_counts <- system.file("extdata/vignette_counts.txt", package = "regsplice")
data <- read.table(file_counts, header = TRUE, sep = "\\t", stringsAsFactors = FALSE)
head(data)

counts <- data[, 2:7]
tbl_exons <- table(sapply(strsplit(data$exon, ":"), function(s) s[[1]]))
gene_IDs <- names(tbl_exons)
n_exons <- unname(tbl_exons)
condition <- rep(c("untreated", "treated"), each = 3)

res <- regsplice(counts, gene_IDs, n_exons, condition)

summary_table(res)

}
\seealso{
\code{\linkS4class{RegspliceData}} \code{\linkS4class{RegspliceResults}}
  \code{\link{initialize_results}} \code{\link{filter_zeros}}
  \code{\link{filter_low_counts}} \code{\link{run_normalization}}
  \code{\link{run_voom}} \code{\link{create_design_matrix}}
  \code{\link{fit_reg_multiple}} \code{\link{fit_null_multiple}}
  \code{\link{fit_full_multiple}} \code{\link{LR_tests}} \code{\link{summary_table}}
}

