% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrix.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Create design matrix.}
\usage{
create_design_matrix(condition, n_exons)
}
\arguments{
\item{condition}{Vector containing the biological conditions for each sample 
(character or numeric vector, or factor).}

\item{n_exons}{Number of exons in the gene (integer).}
}
\value{
Returns a design matrix for the gene, in the format required by the
  \emph{regsplice} model fitting functions.
}
\description{
Creates a design matrix for a single gene.
}
\details{
Creates a design matrix for a single gene in the format required by the
\emph{regsplice} fitting functions. Required inputs are the conditions for each
biological sample, and the number of exons in the gene.

The fitting functions in subsequent steps will call this function once for each gene.

Notes:

\itemize{
\item The design matrix does not contain any main effect terms for the conditions, 
since these are absorbed within the main effect terms for the samples.

\item The design matrix does not include an intercept column, since it is simpler to 
let the model fitting functions add it back later.
}
}
\examples{
condition <- rep(c(0, 1), each = 3)
n_exons <- 10
create_design_matrix(condition, n_exons)

}

