% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions_single_gene.R
\name{fitRegModelSingle}
\alias{fitGLMSingle}
\alias{fitNullModelSingle}
\alias{fitRegModelSingle}
\title{Fit model for a single gene.}
\usage{
fitRegModelSingle(X = NULL, Y, weights = NULL, group = NULL,
  nexons = NULL, alpha = 1, lambda_choice = c("lambda.min", "lambda.1se"),
  ...)

fitGLMSingle(X = NULL, Y, weights = NULL, group = NULL, nexons = NULL,
  ...)

fitNullModelSingle(X = NULL, Y, weights = NULL, group = NULL,
  nexons = NULL, ...)
}
\description{
Internal functions to fit models for a single gene.
}
\details{
\code{fitRegModelSingle} fits a regularized model.

\code{fitGLMSingle} fits a full GLM containing all interaction terms.

\code{fitNullModelSingle} fits a null model with no interaction terms.

Each function fits a model for a single gene only. These are intended as internal
functions. In most cases, the user will only need to access the external functions
\code{\link{fitRegModel}}, \code{\link{fitGLM}}, and \code{\link{fitNullModel}}, which
can fit models for multiple genes using parallelization.

The external functions also call the internal function \code{\link{fitParallel}}.

See documentation for \code{\link{fitRegModel}}, \code{\link{fitGLM}}, and
\code{\link{fitNullModel}} for more details, including parameters and return values.
}
\seealso{
\code{\link{fitRegModel}} \code{\link{fitGLM}} \code{\link{fitNullModel}} 
  \code{\link{fitParallel}} \code{\link{lrTest}} \code{\link[glmnet]{glmnet}} 
  \code{\link[glmnet]{cv.glmnet}} \code{\link[stats]{glm}}
}

