% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RegspliceResults.R
\docType{class}
\name{RegspliceResults-class}
\alias{RegspliceResults}
\alias{RegspliceResults-class}
\title{RegspliceResults objects.}
\usage{
RegspliceResults(gene_IDs)
}
\arguments{
\item{gene_IDs}{Gene identifiers or names (character vector).}
}
\value{
Returns an empty \code{RegspliceResults} object.
}
\description{
\code{RegspliceResults} objects contain the results of a \code{regsplice} analysis.
}
\details{
\code{RegspliceResults} objects are created with the constructor function
\code{RegspliceResults()}, which requires the gene IDs as an argument.

Once created, \code{RegspliceResults} objects are then populated using the functions 
\code{\link{fit_reg_multiple}}, \code{\link{fit_null_multiple}},
\code{\link{fit_full_multiple}}, and \code{\link{LR_tests}}.

The function \code{\link{summary_table}} can be used to display a summary table of the
results.
}
\section{Fields}{

\describe{
\item{\code{gene_IDs}}{Gene identifiers or names (character vector).}

\item{\code{fit_reg_models}}{Fitted regularized (lasso) model objects from 
\code{\link{fit_reg_multiple}}.}

\item{\code{fit_reg_dev}}{Deviance of fitted regularized (lasso) models from 
\code{\link{fit_reg_multiple}}.}

\item{\code{fit_reg_df}}{Degrees of freedom of fitted regularized (lasso) models from 
\code{\link{fit_reg_multiple}}.}

\item{\code{fit_null_models}}{Fitted null model objects from \code{\link{fit_null_multiple}}.}

\item{\code{fit_null_dev}}{Deviance of fitted null models from 
\code{\link{fit_null_multiple}}.}

\item{\code{fit_null_df}}{Degrees of freedom of fitted null models from 
\code{\link{fit_null_multiple}}.}

\item{\code{fit_full_models}}{Fitted full model objects from \code{\link{fit_full_multiple}}.}

\item{\code{fit_full_dev}}{Deviance of fitted full models from 
\code{\link{fit_full_multiple}}.}

\item{\code{fit_full_df}}{Degrees of freedom of fitted full models from 
\code{\link{fit_full_multiple}}.}

\item{\code{p_val}}{Raw p-values (numeric vector).}

\item{\code{p_adj}}{Multiple testing adjusted p-values (Benjamini-Hochberg false discovery 
rates, FDR).}

\item{\code{LR_stat}}{Likelihood ratio test statistics.}

\item{\code{df_test}}{Degrees of freedom of likelihood ratio tests.}
}}
\examples{
# initialize RegspliceResults object
gene_IDs <- paste0("gene", 1:5)
RegspliceResults(gene_IDs)

}
\seealso{
\code{\link{fit_reg_multiple}} \code{\link{fit_null_multiple}}
  \code{\link{fit_full_multiple}} \code{\link{LR_tests}} \code{\link{summary_table}}
}

